# spec/openapi.yaml
openapi: 3.1.0
info:
  title: arw-server
  description: Unified ARW server API surface (headless-first).
  license:
    name: MIT OR Apache-2.0
    identifier: MIT OR Apache-2.0
  version: 0.1.4
paths:
...
  /admin/chat:
    get:
      tags:
      - Admin/Chat
      operationId: chat_history_doc
      responses:
        '200':
          description: Current chat history
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChatHistory'
        '401':
          description: Unauthorized
      description: 'Chat: GET /admin/chat.'
  /admin/chat/clear:
    post:
      tags:
      - Admin/Chat
      operationId: chat_clear_doc
      responses:
        '200':
          description: Cleared
          content:
            application/json:
              schema: {}
        '401':
          description: Unauthorized
      description: 'Chat: POST /admin/chat/clear.'
  /admin/chat/send:
    post:
      tags:
      - Admin/Chat
      operationId: chat_send_doc
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChatSendReq'
        required: true
      responses:
        '200':
          description: Synthetic reply
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChatSendResp'
        '401':
          description: Unauthorized
      description: 'Chat: POST /admin/chat/send.'
  /admin/chat/status:
    get:
      tags:
      - Admin/Chat
      operationId: chat_status_doc
      parameters:
      - name: probe
        in: query
        description: Trigger latency probe
        required: false
        schema:
          type: boolean
      responses:
        '200':
          description: Backend status
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChatStatusResp'
        '401':
          description: Unauthorized
      description: 'Chat: GET /admin/chat/status.'
...
  /events:
    get:
      tags:
      - Events
      summary: "Server\u2011Sent Events stream of envelopes."
      description: "Server\u2011Sent Events stream of envelopes; supports replay and\
        \ prefix filters."
      operationId: events_sse_doc
      parameters:
      - name: after
        in: query
        description: Resume after id or Last-Event-ID header
        required: false
        schema:
          type: integer
          format: int64
      - name: replay
        in: query
        description: Replay the last N events (when after not set)
        required: false
        schema:
          type: integer
          minimum: 0
      - name: prefix
        in: query
        description: CSV of event kind prefixes to include
        required: false
        schema:
          type: string
      responses:
        '200':
          description: SSE stream of events
          content:
            text/event-stream: {}
        '501':
          description: Kernel disabled
          content:
            application/json:
              schema: {}
...
components:
  schemas:
...
    ChatHistory:
      type: object
      required:
      - items
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/ChatMessage'
    ChatMessage:
      type: object
      required:
      - id
      - role
      - content
      - ts_ms
      properties:
        content:
          type: string
        id:
          type: string
        role:
          type: string
        ts_ms:
          type: integer
          format: int64
          minimum: 0
    ChatSendReq:
      type: object
      required:
      - prompt
      properties:
        prompt:
          type: string
    ChatSendResp:
      type: object
      required:
      - ok
      - backend
      - reply
      - history
      properties:
        backend:
          type: string
        history:
          type: array
          items:
            $ref: '#/components/schemas/ChatMessage'
        ok:
          type: boolean
        reply:
          $ref: '#/components/schemas/ChatMessage'
    ChatStatusResp:
      type: object
      required:
      - ok
      - backend
      - messages
      properties:
        backend:
          type: string
        latency_ms:
          type:
          - integer
          - 'null'
          format: int64
          minimum: 0
        messages:
          type: integer
          format: int64
          minimum: 0
        ok:
          type: boolean
...